+++ flutter/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Made the builder's add method support default arguments.
+* Added the set method to SBMLBuilder.
+
 ## 0.0.1
 
 * Initial release.